feat: fallback to xrandr
In some cases inxi can't find basic graphics data. In this case let's fallback
to xrandr to complement the data.

In the future we might additionally look into a fallback to wayland-info. But
as a first step the xrandr fallback is sufficient as it is also available on
Wayland via Xwayland.